Refactor nightly CI workflow for DeepSeek-V3.2 performance tests
- Introduced separate performance test scripts for DeepSeek-V3.2 basic and MTP variants, enhancing test coverage on MI35x.
- Updated the nightly CI workflow to include these new performance jobs, replacing the previous combined performance test.
- Adjusted accuracy test documentation to reflect changes in configurations and removed outdated comments.
- Improved job timeout settings and streamlined the reporting process for better clarity in results.
