+++ .github/workflows/nightly-test-amd.yml
@@ -40,7 +40,8 @@ on:
           - 'nightly-8-gpu-mi35x-grok2'
           - 'nightly-8-gpu-mi35x-deepseek-r1-mxfp4'
           - 'nightly-accuracy-8-gpu-mi35x-deepseek-v32'
-          - 'nightly-perf-8-gpu-mi35x-deepseek-v32'
+          - 'nightly-perf-8-gpu-mi35x-deepseek-v32-basic'
+          - 'nightly-perf-8-gpu-mi35x-deepseek-v32-mtp'
   workflow_call:
     inputs:
       ref:
@@ -643,9 +644,9 @@ jobs:
           echo "$(<github_summary.md )" >> $GITHUB_STEP_SUMMARY || true
           exit ${TEST_EXIT_CODE:-0}
 
-  # MI35x 8-GPU DeepSeek-V3.2 Performance Test
-  nightly-perf-8-gpu-mi35x-deepseek-v32:
-    if: (github.repository == 'sgl-project/sglang' || github.event_name == 'pull_request') && (inputs.job_filter == '' || inputs.job_filter == 'all' || inputs.job_filter == 'nightly-perf-8-gpu-mi35x-deepseek-v32')
+  # MI35x 8-GPU DeepSeek-V3.2 Performance Test (Basic)
+  nightly-perf-8-gpu-mi35x-deepseek-v32-basic:
+    if: (github.repository == 'sgl-project/sglang' || github.event_name == 'pull_request') && (inputs.job_filter == '' || inputs.job_filter == 'all' || inputs.job_filter == 'nightly-perf-8-gpu-mi35x-deepseek-v32-basic')
     runs-on: linux-mi35x-gpu-8
     steps:
       - name: Checkout code
@@ -666,13 +667,46 @@ jobs:
           # Install tabulate for run_suite.py (missing in MI35x container)
           bash scripts/ci/amd_ci_exec.sh pip install tabulate
 
-      - name: Performance Test MI35x (8-GPU DeepSeek-V3.2)
-        timeout-minutes: 300
+      - name: Performance Test MI35x (8-GPU DeepSeek-V3.2 Basic)
+        timeout-minutes: 150
         run: |
           > github_summary.md  # Clear summary file
           bash scripts/ci/amd_ci_exec.sh -w /sglang-checkout/test \
             -e GITHUB_STEP_SUMMARY="/sglang-checkout/github_summary.md" \
-            python3 run_suite.py --hw amd --suite nightly-perf-8-gpu-mi35x-deepseek-v32 --nightly --timeout-per-file 18000 || TEST_EXIT_CODE=$?
+            python3 run_suite.py --hw amd --suite nightly-perf-8-gpu-mi35x-deepseek-v32-basic --nightly --timeout-per-file 5400 || TEST_EXIT_CODE=$?
+          echo "$(<github_summary.md )" >> $GITHUB_STEP_SUMMARY || true
+          exit ${TEST_EXIT_CODE:-0}
+
+  # MI35x 8-GPU DeepSeek-V3.2 Performance Test (MTP)
+  nightly-perf-8-gpu-mi35x-deepseek-v32-mtp:
+    if: (github.repository == 'sgl-project/sglang' || github.event_name == 'pull_request') && (inputs.job_filter == '' || inputs.job_filter == 'all' || inputs.job_filter == 'nightly-perf-8-gpu-mi35x-deepseek-v32-mtp')
+    runs-on: linux-mi35x-gpu-8
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v4
+        with:
+          ref: ${{ inputs.ref || github.ref }}
+
+      - name: Setup docker
+        run: |
+          touch github_summary.md
+          bash scripts/ci/amd_ci_start_container.sh
+        env:
+          GITHUB_WORKSPACE: ${{ github.workspace }}
+
+      - name: Install dependencies
+        run: |
+          bash scripts/ci/amd_ci_install_dependency.sh
+          # Install tabulate for run_suite.py (missing in MI35x container)
+          bash scripts/ci/amd_ci_exec.sh pip install tabulate
+
+      - name: Performance Test MI35x (8-GPU DeepSeek-V3.2 MTP)
+        timeout-minutes: 150
+        run: |
+          > github_summary.md  # Clear summary file
+          bash scripts/ci/amd_ci_exec.sh -w /sglang-checkout/test \
+            -e GITHUB_STEP_SUMMARY="/sglang-checkout/github_summary.md" \
+            python3 run_suite.py --hw amd --suite nightly-perf-8-gpu-mi35x-deepseek-v32-mtp --nightly --timeout-per-file 5400 || TEST_EXIT_CODE=$?
           echo "$(<github_summary.md )" >> $GITHUB_STEP_SUMMARY || true
           exit ${TEST_EXIT_CODE:-0}
 
@@ -700,7 +734,8 @@ jobs:
       - nightly-8-gpu-mi35x-grok2
       - nightly-8-gpu-mi35x-deepseek-r1-mxfp4
       - nightly-accuracy-8-gpu-mi35x-deepseek-v32
-      - nightly-perf-8-gpu-mi35x-deepseek-v32
+      - nightly-perf-8-gpu-mi35x-deepseek-v32-basic
+      - nightly-perf-8-gpu-mi35x-deepseek-v32-mtp
     runs-on: ubuntu-latest
     steps:
       - name: Check if any job failed
